searchdoc/app/requirements.txt
google-cloud-discoveryengine==0.11.10
google-cloud-storage==2.14.0
fsspec==2024.2.0
gcsfs==2024.2.0
python-dotenv==1.0.1
flask==3.0.2
google-cloud==0.34.0
ipython
#Flask
google-cloud
google-cloud-storage==2.14.0
fsspec==2024.2.0
gcsfs==2024.2.0
python-dotenv
Flask-MySQLdb
Flask-SQLAlchemy
pymysql
python-jose 
flask-jwt-extended
fastapi
bcrypt
passlib
phpserialize
cryptography
gunicorn